+++ Assets/Scripts/UI/Menus/Controls/VideoMappingMenu.cs
@@ -1,5 +1,6 @@
 ﻿using UnityEngine;
 using UnityEngine.SceneManagement;
+using VoyagerApp.Networking.Packages.Voyager;
 using VoyagerApp.Utilities;
 using VoyagerApp.Videos;
 
@@ -27,8 +28,11 @@ namespace VoyagerApp.UI.Menus
         private void FpsChanged(int value)
         {
             video.fps = value;
+            var packet = new SetFpsPacket(value);
+
             foreach (var lamp in WorkspaceUtils.Lamps)
-                lamp.SetVideo(video);
+                NetUtils.VoyagerClient.SendPacket(lamp, packet);
+
             mapper.SetFps(value);
         }